% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countOligos.R
\docType{methods}
\name{countOligos}
\alias{countOligos}
\title{Count oligos in each set of fastq files.}
\usage{
countOligos(oligos, fastq.names)
}
\arguments{
\item{oligos}{A data.frame of oligos from readOligos}

\item{fastq.names}{A list of R1 and R2 reads from readFastqFileNames}
}
\value{
A matrix of counts for each oligo (row) and fastq file (column)
}
\description{
Reads in each set of fastq files and counts oligos within
}
\details{
countOligos
Counts up each oligo in each set of R1/R2 fastq files.
}
\examples{
oligo.counts <- countOligos(oligos,fq.files)
}
