% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFastqFileNames.R
\docType{methods}
\name{readFastqFileNames}
\alias{readFastqFileNames}
\title{Read fastq file names}
\usage{
readFastqFileNames(path = ".", pattern)
}
\arguments{
\item{path}{Path to the folder in which to look for fastq files. Default is current working folder.}

\item{pattern}{A pattern in the file names common to all.}
}
\value{
A list containing vectors of R1 and R2 reads
}
\description{
Reads in just the file names of the fastq files needing scanning.
}
\details{
readFastqFileNames
Reads in the names of the fastq files which need to be scanned
}
\examples{
fq.files <- readFastqFileNames(path="."pattern="fastq")
}
