% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOligos.R
\docType{methods}
\name{readOligos}
\alias{readOligos}
\title{Read oligos.}
\usage{
readOligos(file)
}
\arguments{
\item{file}{A file containing the oligos}
}
\value{
A data.frame of oligios.
}
\description{
Reads a comma separated file of oligos where the first column is oligo name, and the second the sequeunce.
}
\details{
readOligos
Reads in the oligos that need to be found in the fastq files.
}
\examples{
oli.ebf <- readOligos("Ebf1_library_oligos.csv")
}
