% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeEndCounts.R, R/readFastqFileNames.R
\docType{methods}
\name{readFastqFileNames}
\alias{readFastqFileNames}
\alias{mergeEndCounts}
\title{Merge counts}
\usage{
mergeEndCounts(counts)

readFastqFileNames(path = ".", pattern)
}
\arguments{
\item{counts}{A table from the output of countOligos}

\item{path}{Path to the folder in which to look for fastq files. Default is current working folder.}

\item{pattern}{A pattern in the file names common to all.}
}
\value{
A table of merged counts.

A list containing vectors of R1 and R2 reads
}
\description{
Takes the output from countOligos and merges the R1 and R2 reads

Reads in just the file names of the fastq files needing scanning.
}
\details{
mergeEndCounts
Merge the counts from R1 and R2 and F and R oligos

readFastqFileNames
Reads in the names of the fastq files which need to be scanned
}
\examples{
mergedOligoCounts <- mergeEndCounts(oligo.counts)
fq.files <- readFastqFileNames(path="."pattern="fastq")
}
